backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
groq
google-generativeai
langchain
langchain-community
chromadb
sentence-transformers